record_newとrecord_editのstimulusコントローラーを統合
