+++ app/views/records/new.html.erb
@@ -1,18 +1,18 @@
-<div data-controller="record-new">
+<div data-controller="record">
   <div class="container mx-auto py-8 px-3 md:px-0">
     <div class="w-full max-w-md mx-auto bg-base-200 rounded-lg p-6 md:p-8">
       <h1 class="text-2xl font-bold text-center mb-8"><%= t('records.new.create_record') %></h1>
 
-      <%= form_with model: @record, local: true, html: { id: 'record_form', data: { record_new_target: 'form' } } do |form| %>
+      <%= form_with model: @record, local: true, html: { id: 'record_form', data: { record_target: 'form' } } do |form| %>
         <div>
           <label for="item_search" class="block text-accent font-bold mb-2"><%= t('records.new.search_items') %></label>
-          <input data-record-new-target="searchInput" type="text" id="item_search" name="item_search" class="appearance-none border rounded w-full bg-white py-2 px-3 text-base-100 leading-tight focus:outline-none focus:shadow-outline" placeholder="<%= t('records.new.title_artist_label') %>">
+          <input data-record-target="searchInput" type="text" id="item_search" name="item_search" class="appearance-none border rounded w-full bg-white py-2 px-3 text-base-100 leading-tight focus:outline-none focus:shadow-outline" placeholder="<%= t('records.new.title_artist_label') %>">
         </div>
 
-        <div data-record-new-target="resultsContainer" id="search_results" class="max-h-48 overflow-y-auto rounded bg-white text-base-100 text-sm md:text-base p-2 mb-4"></div>
+        <div data-record-target="resultsContainer" id="search_results" class="max-h-48 overflow-y-auto rounded bg-white text-base-100 text-sm md:text-base p-2 mb-4"></div>
 
         <h2 class="font-bold mb-4 mt-4"><%= t('records.daily_records.listened_records') %></h2>
-        <div data-record-new-target="selectedItems" id="selected_items" class="mb-4 mt-4 text-sm md:text-base" ></div>
+        <div data-record-target="selectedItems" id="selected_items" class="mb-4 mt-4 text-sm md:text-base" ></div>
 
         <div class="mb-4">
           <%= form.label t('records.new.comment'), class: 'block text-accent font-bold mb-2' %>
